+++ src/si/papirservis/XLSCreateProdajaServlet.java
@@ -144,7 +144,7 @@ public class XLSCreateProdajaServlet extends InitServlet implements Servlet {
 		// write data
 		r = s.createRow(rownum);
 		
-		for (int n=0; n<rowNames.length;n++) {
+		for (short n=0; n<rowNames.length;n++) {
 			c = r.createCell(n);
 
 			HSSFCellStyle cs = wb.createCellStyle();
@@ -213,7 +213,7 @@ public class XLSCreateProdajaServlet extends InitServlet implements Servlet {
 		//Object[] mun = (Object[]) municipality.keySet().toArray();
 		// write header
 		r = s.createRow(rownum);
-		for (int n=0; n<rowNames.length;n++) {
+		for (short n=0; n<rowNames.length;n++) {
 			c = r.createCell(n);
 			c.setCellStyle(cs);
 			if (rowNames[n].equals(""))
